requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
EasyProcess==1.1
entrypoint2==1.1
fcm-django==2.0.0
firebase-admin==6.2.0
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
# MouseInfo==0.1.3
msgpack==0.6.2
mss==9.0.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
# pywhatkit==5.4
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==6.7.0
stripe-Webhook==0.1.8
# Twisted==23.8.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
Werkzeug==3.0.0
wikipedia==1.4.0
zope.interface==6.1
